requirements.txt
fastapi==0.75.1
numpy==1.21.5
pandas==1.3.5
requests==2.27.1
uvicorn==0.17.6
fsspec==2022.3.0
faiss-cpu==1.7
html2text==2020.1.16
google-cloud-translate==3.8.4
firebase-admin==6.0.1
ftfy==6.1.1
regex==2022.9.13
tqdm==4.64.1
torch==1.12.1
torchvision==0.13.1
google-cloud-retail==1.13.0
google-auth==2.15.0